【安全】修复Summary Communication模块CleanCode
# message auto-generated for no-merge-commit merge:
merge cleancode_****** into master

【安全】修复Summary Communication模块CleanCode

Created-by: insightDeveloper

Match-id-2eb40d8b9b0c99da20f9e0e6fa5992092c2f09c9

+++ server/src/modules/communication/parser/ClusterFileParser.cpp
@@ -478,7 +478,7 @@ std::string ClusterFileParser::GetStrValue(std::map<std::string, size_t> &dataMa
     if (dataMap.find(key) == dataMap.end()) {
         return "";
     }
-    int index = dataMap[key];
+    size_t index = dataMap[key];
     return tokens[index];
 }
